% Generated by roxygen2 (4.0.1): do not edit by hand
\name{memory.size}
\alias{memory.limit}
\alias{memory.size}
\title{Report on Memory Allocation}
\usage{
memory.size(max = FALSE)

memory.limit(size = NA)
}
\arguments{
\item{max}{logical. If true the maximum amount of memory obtained from the
OS is reported, otherwise the amount currently in use.}

\item{size}{numeric. If \code{NA} report the memory size, otherwise request
a new limit, in Mb.}
}
\value{
Size in bytes
}
\description{
\code{memory.size} and \code{memory.limit} are used to manage the total
memory allocation.  On POSIX systems, the \code{getrlimit64(2)}
and  \code{setrlimit64(2)} APIs are used.  On Windows, these calls are
forwarded to the corresponding functions in the \code{utils} package.
}
\seealso{
\link[base]{Memory-limits} for other limits.
}

