% Generated by roxygen2 (4.0.1): do not edit by hand
\name{memory_limit}
\alias{memory_limit}
\title{Report on Memory Allocation}
\usage{
memory_limit(size = NA)
}
\arguments{
\item{size}{numeric. If \code{NA} report the memory size, otherwise request
a new limit, in MiB.}
}
\value{
A named vector with the (new) soft and hard limits in MiB.
}
\description{
This function is used to manage the total memory allocation.  On POSIX
systems, the \code{getrlimit64(2)} and  \code{setrlimit64(2)} APIs are used.
On Windows, these calls are
forwarded to \code{utils::\link[utils]{memory.limit}}.
}
\seealso{
\link[base]{Memory-limits} for other limits.
}

