% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{ulimit-package}
\alias{ulimit}
\alias{ulimit-package}
\title{Title}
\description{
Description
}
\details{
\tabular{ll}{
  Package: \tab ulimit\cr
  Type: \tab Package\cr
  Version: \tab 0.0-2\cr
  Date: \tab 2014-06-29\cr
  License: \tab GPL v3\cr
}
}
\author{
Kirill Müller
}
\keyword{package}

