% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlimit.R
\name{memory_limit}
\alias{memory_limit}
\title{Query and set memory limits}
\usage{
memory_limit(size = NA)
}
\arguments{
\item{size}{numeric. If \code{NA} report the memory size, otherwise request
a new limit, in MiB.}
}
\value{
A named vector with the (new) soft and hard limits in MiB.
}
\description{
This function is used to manage the total memory allocation
for the current session and for child processes.  On POSIX
systems, the \href{http://linux.die.net/man/2/getrlimit}{\code{getrlimit64(2)} and  \code{setrlimit64(2)} APIs}
with resource \code{RLIMIT_AS} are used.
Support for Windows might become available in the future; in the meantime,
use \code{utils::\link[utils]{memory.limit}}.
}
\seealso{
\link[base]{Memory-limits} for other limits.
}

