% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ulimit-package.R
\docType{package}
\name{ulimit-package}
\alias{ulimit}
\alias{ulimit-package}
\title{Enforce memory limits on POSIX systems}
\description{
This package provides a way to enforce memory limits for a
  running R process on POSIX systems.  The function \code{\link{memory_limit}}
  shows and optionally sets the limits on main memory.
}
\examples{
\dontrun{
# Limit available memory to 6000 GiB
memory_limit(6000)}
}
\author{
Kirill Müller
}
\keyword{package}

